+++ locallib.php
@@ -26,6 +26,7 @@ define('COURSE_MERGE_DEPTH_UNLIMITED', 0);
 define('COURSE_MERGE_DEPTH_SAME_CATEGORY', 1);
 define('COURSE_MERGE_DEPTH_SAME_PARENT', 2);
 define('COURSE_MERGE_DEFAULT_CATEGORY', -1);
+define('COURSE_MERGE_USE_DEFAULT_GROUPMODE', -1);
 
 /**
  * Get the advanced fields for the course merge tool.